[Foundation] Move the NSHttpCookieExtensions out of the handler file.
