+++ src/Makefile
@@ -134,11 +134,13 @@ IOS_EXTRA_SOURCES = \
 
 IOS_DOTNET_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 
 IOS_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 
@@ -515,6 +517,7 @@ MAC_CFNETWORK_SOURCES = \
 
 MAC_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 
@@ -829,6 +832,7 @@ WATCHOS_EXTRA_CORE_SOURCES = \
 
 WATCHOS_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 
 WATCHOS_CORE_SOURCES +=           \
     $(WATCHOS_EXTRA_CORE_SOURCES) \
@@ -1121,12 +1125,14 @@ TVOS_CORE_SOURCES +=           \
 
 TVOS_DOTNET_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 	$(SHARED_DESIGNER_CS) \
 
 TVOS_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 	$(SHARED_DESIGNER_CS) \
@@ -1338,12 +1344,14 @@ MACCATALYST_DOTNET_CORE_SOURCES += \
 
 MACCATALYST_DOTNET_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 	$(SHARED_DESIGNER_CS) \
 
 MACCATALYST_HTTP_SOURCES = \
 	Foundation/NSUrlSessionHandler.cs \
+	Foundation/NSUrlSessionHandler/NSHttpCookieExtensions.cs \
 	System.Net.Http/CFContentStream.cs \
 	System.Net.Http/CFNetworkHandler.cs \
 	$(SHARED_DESIGNER_CS) \
